refactor: get-blog-info 로 파일 이름 변경 (#447)
