+++ src/app/series/page.tsx
@@ -3,7 +3,7 @@ import { Metadata } from 'next';
 import phrases from 'data/phrases';
 import siteMetadata from 'data/siteMetadata.mjs';
 
-import { getAllSeries } from '@/lib/getBlogInfo.mjs';
+import { getAllSeries } from '@/lib/get-blog-info.mjs';
 
 import SeriesList from '@/components/card-and-list/series-list';
